refactor: move geometries into sererate file
